update README, multiple actions, and structure
