+++ Dalaran/parser/__init__.py
@@ -1,1 +1,1 @@
-from .main import *
+from .cards import *
