+++ README.md
@@ -4,9 +4,7 @@
 
 Basic Lexer and Parser for Hearthstone Simulations.
 
-Built in the hope of creating an easy way to add custom cards to a simulation and to learn parsing and lexing a "language".
-
-Currently in the stage of building out the lexer to tokenize all current cards correctly and getting basic parsing working.
+Effectively, this project aims to provide a tool that can "read" Hearthstone cards and generate playable versions of them from their text.
 
 In it's current state, the lexer is able to lex cards from the Basic Set or cards that have the same vocabulary. It can parse card's abilities and basic actions like drawing cards into a form usable in [fireplace](https://github.com/jleclanche/fireplace), a Python implementation of Hearthstone.
 
@@ -31,6 +29,23 @@ Install Dalaran:
 pip3 install .
 ```
 
-IN PROGRESS 
+Open up a shell for `python3.6` or `python3.7`, now we're going to make a very weak, basic spell:
+```python
+>>> from Dalaran.lexer import lex, token_exprs
+>>> from Dalaran.parser import parse_card, register_card
+>>> from Dalaran.utils import prepare_game
+>>> tokens = lex('Deal 10 Damage. Draw 5 Cards.', token_exprs) # tokenize the text
+>>> card = parse_card('Weak Spell', 'Spell', 1, 'Mage', tokens) # create the card
+>>> register_card(card)
+>>> game = prepare_game()
+>>> weak_spell = game.player1.give(card.__name__) # give the card to the first player
+>>> print(len(game.player1.hand)) # 5
+>>> print(game.player1.opponent.hero.health) # 30
+>>> weak_spell.play(target=game.player1.opponent.hero) # play the spell
+>>> print(len(game.player1.hand)) # 5 + 5 - 1 = 9
+>>> print(game.player1.opponent.hero.health) # 30 - 10 = 20
+```
+
+And that's how you can make and play with your custom hearthstone cards!
 
 Feel free to help out and contribute!
